+++ gym_sts/envs/utils.py
@@ -80,3 +80,16 @@ def obs_value(obs: Observation) -> float:
     value = float(obs.persistent_state.floor)
     value += obs.persistent_state.hp / 100
     return value
+
+
+def single_combat_value(obs: Observation) -> float:
+    max_hp = sum(e.max_hp for e in obs.combat_state.enemies)
+    enemy_hp = sum(e.current_hp for e in obs.combat_state.enemies)
+
+    self_hp = obs.persistent_state.hp
+
+    if max_hp == 0:
+        enemy_hp = 1
+        max_hp = 1
+
+    return ((max_hp - enemy_hp) / max_hp) * 100 + self_hp
